+++ flutter/lib/src/sentry_asset_bundle.dart
@@ -9,7 +9,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter/material.dart';
 import 'package:sentry/sentry.dart';
 
-typedef _Parser<T> = Future<T> Function(String value);
+typedef _StringParser<T> = Future<T> Function(String value);
+typedef _ByteParser<T> = FutureOr<T> Function(ByteData value);
 
 /// An [AssetBundle] which creates automatic performance traces for loading
 /// assets.
@@ -79,7 +80,7 @@ class SentryAssetBundle implements AssetBundle {
   }
 
   @override
-  Future<T> loadStructuredData<T>(String key, _Parser<T> parser) {
+  Future<T> loadStructuredData<T>(String key, _StringParser<T> parser) {
     if (_enableStructuredDataTracing) {
       return _loadStructuredDataWithTracing(key, parser);
     }
@@ -87,7 +88,7 @@ class SentryAssetBundle implements AssetBundle {
   }
 
   Future<T> _loadStructuredDataWithTracing<T>(
-      String key, _Parser<T> parser) async {
+      String key, _StringParser<T> parser) async {
     final span = _hub.getSpan()?.startChild(
           'file.read',
           description: 'AssetBundle.loadStructuredData<$T>: ${_fileName(key)}',
@@ -103,6 +104,46 @@ class SentryAssetBundle implements AssetBundle {
       final data = await _bundle.loadStructuredData(
         key,
         (value) async => await _wrapParsing(parser, value, key, span),
+      );
+      span?.status = SpanStatus.ok();
+      completer.complete(data);
+    }, (exception, stackTrace) {
+      completer.completeError(exception, stackTrace);
+    });
+
+    T data;
+    try {
+      data = await completer.future;
+      _setDataLength(data, span);
+      span?.status = const SpanStatus.ok();
+    } catch (e) {
+      span?.throwable = e;
+      span?.status = const SpanStatus.internalError();
+      rethrow;
+    } finally {
+      await span?.finish();
+    }
+    return data;
+  }
+
+  FutureOr<T> _loadStructuredBinaryDataWithTracing<T>(
+      String key, _ByteParser<T> parser) async {
+    final span = _hub.getSpan()?.startChild(
+          'file.read',
+          description:
+              'AssetBundle.loadStructuredBinaryData<$T>: ${_fileName(key)}',
+        );
+    span?.setData('file.path', key);
+
+    final completer = Completer<T>();
+
+    // This future is intentionally not awaited. Otherwise we deadlock with
+    // the completer.
+    // ignore: unawaited_futures
+    runZonedGuarded(() async {
+      final data = await _loadStructuredBinaryDataWrapper(
+        key,
+        (value) async => await _wrapBinaryParsing(parser, value, key, span),
       );
       span?.status = SpanStatus.ok();
       completer.complete(data);
@@ -236,7 +277,7 @@ class SentryAssetBundle implements AssetBundle {
   }
 
   static Future<T> _wrapParsing<T>(
-    _Parser<T> parser,
+    _StringParser<T> parser,
     String value,
     String key,
     ISentrySpan? outerSpan,
@@ -258,5 +299,60 @@ class SentryAssetBundle implements AssetBundle {
     }
 
     return data;
+  }
+
+  static FutureOr<T> _wrapBinaryParsing<T>(
+    _ByteParser<T> parser,
+    ByteData value,
+    String key,
+    ISentrySpan? outerSpan,
+  ) async {
+    final span = outerSpan?.startChild(
+      'serialize.file.read',
+      description: 'parsing "$key" to "$T"',
+    );
+    T data;
+    try {
+      data = await parser(value);
+      span?.status = const SpanStatus.ok();
+    } catch (e) {
+      span?.throwable = e;
+      span?.status = const SpanStatus.internalError();
+      rethrow;
+    } finally {
+      await span?.finish();
+    }
+
+    return data;
+  }
+
+  @override
+  // ignore: override_on_non_overriding_member
+  Future<T> loadStructuredBinaryData<T>(
+    String key,
+    FutureOr<T> Function(ByteData data) parser,
+  ) async {
+    if (_enableStructuredDataTracing) {
+      return _loadStructuredBinaryDataWithTracing<T>(key, parser);
+    }
+
+    return _loadStructuredBinaryDataWrapper<T>(key, parser);
+  }
+
+  // helper method to have a "typesafe" method
+  Future<T> _loadStructuredBinaryDataWrapper<T>(
+    String key,
+    FutureOr<T> Function(ByteData data) parser,
+  ) async {
+    // The loadStructuredBinaryData method exists as of Flutter greater than 3.8
+    // Previous versions don't have it, but later versions do.
+    // We can't use `extends` in order to provide this method because this is
+    // a wrapper and thus the method call must be forwarded.
+    // On Flutter versions <=3.8 we can't forward this call.
+    // On later version the call gets correctly forwarded.
+    // The error doesn't need to handled since it can't be called on earlier versions,
+    // and it's correctly forwarded on later versions.
+    return (_bundle as dynamic).loadStructuredBinaryData<T>(key, parser)
+        as Future<T>;
   }
 }